feat: implement CheckoutPage with payment processing and success handling
