+++ src/app/checkout/[bid]/page.tsx
@@ -1,7 +1,7 @@
 import { getServerSession } from "next-auth";
 import { authOptions } from "@/app/api/auth/[...nextauth]/authOptions";
 import getUserProfile from "@/libs/Auth/getUserProfile";
-import CheckoutCard from "@/components/Checkout";
+import CheckoutPage from "@/components/CheckoutPage";
 import { getPayment } from "@/libs/Payment/getPayment";
 
 export default async function Checkout({ params }: { params: { bid: string } }) {
@@ -11,13 +11,11 @@ export default async function Checkout({ params }: { params: { bid: string } }) 
 
     const profile = await getUserProfile(session.user.token);
 
-    const payment = await getPayment(params.bid, session.user.token);
+    const payment: PaymentJsonOne = await getPayment(params.bid, session.user.token);
     
-    console.log(payment.data);
-
     return (
-        <main >
-            <CheckoutCard paymentData={payment.data} />
+        <main>
+            <CheckoutPage token={session.user.token} paymentJson={payment}/>
         </main>
     );
 }
